+++ app/build.gradle.kts
@@ -82,7 +82,7 @@ tasks.jacocoTestCoverageVerification {
             limit {
                 counter = "LINE"
                 value = "COVEREDRATIO"
-                minimum = BigDecimal.valueOf(0.4)
+                minimum = BigDecimal.valueOf(0.2)
             }
         }
     }
